Workaround fix for a subscribing client to ack the publish.
When a client subscribes to a topic, the broker sends it the messages
that are published to the topic. For QOS 1, the client needs to
acknowledge that it received the published message. If it doesn't, the
broker would send that published message again.

It seems that there was an issue with client disconnecting too quickly
and not acknowledging the publish, which resulted in broker publishing
the message again causing the tests to fail. For now, this workaround
adds a sleep to give a chance to client to acknowledge before
disconnecting. Long term fix is pending.